usb: dwc3-msm: Relax PM wakelock if PM resume from B_SUSPEND state
The driver holds PM wakelock if PM resume is called. If the
device was in B_SUSPEND(peripheral mode bus suspend) when
suspended earlier, then we will not be releasing the wakelock
from sm_work. Do so in order to allow system suspend to happen
again.

Change-Id: I07396a9ea89b2cdc7683d818c589d2e29bcddfe0
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>

Conflicts:
	drivers/usb/dwc3/dwc3-msm.c

Conflic in (c) year. Missing a previous commit to bridge the gap,
though that one I think will not fly on this older BSP tag.